Biocare/requirements.txt
asgiref==3.11.1
django==6.0.2
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
pyjwt==2.11.0
sqlparse==0.5.5
tzdata==2025.3
